modular init/ueventd config file generation logic
This patch allows different products to share parts of their init.rc
and ueventd.rc configuration.

This is done by splitting the configuration into separate files
together with a build time logic to assemble the target configuration.
It was created to avoid configuration duplication for all devices in the
P1000 family. For now it is disabled for the product p1c.

The following variables are used to build the final .rc files:

TARGET_INITRC_IMPORT
adds "import lines" to the init.<device>.rc file at build time. The to
be imported files have to be copied separately via PRODUCT_COPY_FILES.

TARGET_INITRC_FILES
defines the init.rc files that are joined together to create
init.<device>.rc

TARGET_UEVENTDRC_FILES
defines the ueventd.rc src files that are joined together to create
ueventd.<device>.rc

Change-Id: I2785836445799620dfe89b08256f1030a5814f67