Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

As stated previously, this breaks all video playback via tv out over A/V on p1c. If this truly is not the case with the p1, then a better solution needs to be implemented, rather than just removing the code.

Patch-set: 1