Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

if it's set to anything other than ASSIST, then the hardware key remapping feature will not work. It can be remapped to 'in app search' using that feature if a user wants. ASSIST is the default for crespo/etc.

Patch-set: 1